Moved all the pages into pages directory.
